In the middle of adding tests

+++ src/main/java/com/rover/interview/bgou/tables/ReviewRepository.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 
 @CrossOrigin
 public interface ReviewRepository extends CrudRepository<Review, Long> {
-    int countReviewsBySitterEmail(String email);
+    long countReviewsBySitterEmail(String email);
 
     @Query(value = "SELECT sum(rating) as totalRating FROM review WHERE sitter_email = ?1",
            nativeQuery = true)
